refactor(model): `Lotto`의 패키지 경로 수정
